Struct channel::libra_channel::Sender

pub struct Sender<K: Eq + Hash + Clone, M> { /* fields omitted */ }

The sending end of the libra_channel.

Methods

impl<K: Eq + Hash + Clone, M> Sender<K, M>

pub fn push(&mut self, key: K, message: M) -> Result<()>

This adds the message into the internal queue data structure. This is a synchronous call.

pub fn push_with_feedback(
&mut self,
key: K,
message: M,
status_ch: Option<Sender<ElementStatus<M>>>
) -> Result<()>

Same as push, but this function also accepts a oneshot::Sender over which the sender can be notified when the message eventually gets delivered or dropped.