Crate channel

Provides an mpsc (multi-producer single-consumer) channel wrapped in an IntGauge

Modules

libra_channel
libra_channel provides an mpsc channel which has two ends libra_channel::Receiver and libra_channel::Sender similar to existing mpsc data structures. What makes it different from existing mpsc channels is that we have full control over how the internal queueing in the channel happens and how we schedule messages to be sent out from this channel. Internally, it uses the PerKeyQueue to store messages

message_queues

Structs

Receiver
Similar to mpsc::Receiver, but with an IntGauge

Sender
Similar to mpsc::Sender, but with an IntGauge

WithEntryTimestamp
Wrapper around a value with an entry timestamp It is used to measure the time waiting in the mpsc::channel.

Statics

TEST_COUNTER

Functions

new
Similar to mpsc::channel, new creates a pair of Sender and Receiver

new_test

new_test_with_timeout

new_with_timeout